Add accountID header for account operations

+++ server/src/main/java/com/example/vvps/controller/AccountController.java
@@ -5,6 +5,7 @@ import com.example.vvps.domain.AccountCreationParameters;
 import com.example.vvps.domain.AccountCredentials;
 import com.example.vvps.repository.AccountRepository;
 import com.example.vvps.service.AccountService;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -30,7 +31,8 @@ public class AccountController {
     }
 
     @PostMapping("/register")
-    public ResponseEntity<Account> createAccount(@RequestBody AccountCreationParameters accountCreationParameters) {
+    public ResponseEntity<Account> createAccount(@RequestBody AccountCreationParameters accountCreationParameters,
+                                                 @RequestHeader("account_id") String id) {
         Account account = Account.builder()
                 .name(accountCreationParameters.getName())
                 .password(accountCreationParameters.getPassword())
@@ -40,18 +42,29 @@ public class AccountController {
     }
 
     @GetMapping("/all")
-    public ResponseEntity<List<Account>> getAll() {
+    public ResponseEntity<List<Account>> getAll(@RequestHeader("account_id") String id) {
+        if (!accountRepository.getIsAdminById(UUID.fromString(id))) {
+            return ResponseEntity.status(403).build();
+        }
         return ResponseEntity.ok(accountRepository.findAll());
     }
 
 
     @GetMapping("/{id}")
-    public ResponseEntity<Account> getById(@PathVariable String id) {
+    public ResponseEntity<Account> getById(@PathVariable String id,
+                                           @RequestHeader("account_id") String loggedAccountId) {
+        if (!accountRepository.getIsAdminById(UUID.fromString(loggedAccountId))) {
+            return ResponseEntity.status(403).build();
+        }
         return ResponseEntity.of(accountRepository.findById(UUID.fromString(id)));
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteById(@PathVariable String id) {
+    public ResponseEntity<Void> deleteById(@PathVariable String id,
+                                           @RequestHeader("account_id") String loggedAccountId) {
+        if (!accountRepository.getIsAdminById(UUID.fromString(loggedAccountId))) {
+            return ResponseEntity.status(403).build();
+        }
         accountService.deleteById(id);
         return ResponseEntity.status(202).build();
     }
